requirements.txt
asgiref==3.8.1
beautifulsoup4==4.11.2
blinker==1.5
certifi==2025.6.15
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.2
cryptography==45.0.5
diff-match-patch==20241021
distro==1.8.0
Django==5.0
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.9
django-axes==8.0.0
django-import-export==4.3.9
django-session-timeout==0.1.0
easypost==10.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.10
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lxml==4.9.2
oauthlib==3.2.2
paypalrestsdk==1.13.3
pillow==11.2.1
pycparser==2.22
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==25.1.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.17
PyQt5-sip==12.11.1
python-debian==0.1.49
python-decouple==3.8
python-docx==0.8.11
pytz==2025.2
pyxattr==0.8.1
pyxdg==0.28
requests==2.32.4
six==1.17.0
soupsieve==2.3.2
sqlparse==0.5.3
tablib==3.8.0
tornado==6.2
typing_extensions==4.14.1
urllib3==2.5.0
wadllib==1.3.6
webencodings==0.5.1
xdg==5.0.0
